fix: remove unused types
