+++ src/types/api/specialty.ts
@@ -1,16 +1,8 @@
-import type { Specialty } from '../common';
-
 export interface UserSpecialty {
   id: string;
   user_id: string;
   specialty_id: string;
   specialty_name: string;
-}
-
-export interface SpecialtyResponse {
-  success: boolean;
-  message: string;
-  data: Specialty[];
 }
 
 export interface UserSpecialtyResponse {
